Fix regionDefinitions typo in pageDesignerHelper.js
